requirements_multimodal.txt
# LOGODETH Multimodal API Requirements
# Simplified stack for AI-powered logo recognition

# Web Framework & API
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6  # For file uploads

# AI/ML APIs
openai==1.3.0
anthropic==0.7.0

# Image Processing
pillow==10.0.1
python-magic==0.4.27  # File type detection

# Caching
redis==5.0.1
hiredis==2.2.3  # C parser for redis (performance)

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3

# HTTP & Async
httpx==0.25.0
aiofiles==23.2.1

# Utilities
python-jose[cryptography]==3.3.0  # For JWT if needed later
passlib[bcrypt]==1.7.4  # For API key hashing if needed

# Monitoring & Logging
loguru==0.7.2
sentry-sdk[fastapi]==1.38.0  # Optional error tracking

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0
mypy==1.7.0

# CORS and Security
fastapi-limiter==0.1.5  # Rate limiting
slowapi==0.1.9  # Alternative rate limiting